+++ docs/components/SuiSelectExampleSpacer.vue
@@ -14,8 +14,9 @@ import SuiInput from "../../src/components/SuiInput.vue";
 
 const selectedItemIndex = ref(0);
 const options = ref([
-    { value: 0, label: 'Option 1' },
-    { value: 1, label: 'Option 2' },
-    { value: 2, label: 'Option 3' },
+    { icon: 'heading', label: 'Heading' },
+    { spacer: true, },
+    { icon: 'align-left', label: 'Text' },
+    { icon: 'double-quotes-l', label: 'Quote' },
 ]);
 </script>